feat(challenge): challenge controller test 기능별 분리
